requirements.txt
pyserial
beautifulsoup4
lxml
requests
Mastodon.py
twitch-chat-irc